backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.23
psycopg[binary]>=3.0.0
transformers>=4.40.0
torch>=2.0.0
python-dotenv==1.0.0
google-generativeai>=0.8.0